+++ packages/ui/index.tsx
@@ -1,2 +1,2 @@
 import * as React from "react";
-export * from "./Button";
+export * from "./base/Button";
